add a test case
